Overhaul of Newtask and EditTask
I added the seperation between type and taskType. The other options like freq and start date are now correctly typed and have fail safes if not correct. Finally, I made it so that when you cant find a task to edit, you just skip instead of continuing on with the code.